Revert ":boom: Svelte!"
