+++ .gitignore
@@ -1,23 +1,42 @@
-node_modules
+# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
-# Output
-.output
-.vercel
-.netlify
-.wrangler
-/.svelte-kit
+# dependencies
+/node_modules
+/.pnp
+.pnp.js
+
+# testing
+/coverage
+
+# next.js
+/.next/
+/out/
+
+# production
 /build
 
-# OS
+# misc
 .DS_Store
-Thumbs.db
+*.pem
 
-# Env
-.env
-.env.*
-!.env.example
-!.env.test
+# debug
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
 
-# Vite
-vite.config.js.timestamp-*
-vite.config.ts.timestamp-*
+# local env files
+.env.local
+.env.development.local
+.env.test.local
+.env.production.local
+
+# vercel
+.vercel
+
+**/*.trace
+**/*.zip
+**/*.tar.gz
+**/*.tgz
+**/*.log
+package-lock.json
+**/*.bun
