--- static/robots.txt
@@ -1,3 +1,0 @@
-# allow crawling everything by default
-User-agent: *
-Disallow:
